move all dynamic acl tests into one spec
since multiple specs might run in parallel in CI, it was causing some
issues. Now the ACLs are updated at the start of the spec and all the
related test cases will run sequentially.
